+++ src/main/java/nextstep/ladder/domain/ConnectPoints.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.domain.vo.Point;
+
 import java.util.Set;
 
 public class ConnectPoints {
