+++ src/main/java/nextstep/ladder/domain/LadderGameSnapshot.java
@@ -1,5 +1,8 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.domain.user.LadderGameUser;
+import nextstep.ladder.domain.vo.Order;
+
 import java.util.Map;
 
 public class LadderGameSnapshot {
