+++ src/main/java/nextstep/ladder/domain/game/LadderResult.java
@@ -1,4 +1,7 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.game;
+
+import nextstep.ladder.domain.LadderGameSnapshot;
+import nextstep.ladder.domain.vo.Point;
 
 import java.util.Map;
