+++ src/main/java/nextstep/ladder/domain/user/LadderGameUser.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.user;
 
 import java.util.Objects;
