+++ src/main/java/nextstep/ladder/domain/user/LadderGameUsers.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.user;
 
 import java.util.Collection;
 import java.util.HashSet;
