+++ src/main/java/nextstep/ladder/domain/vo/Order.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.vo;
 
 
 import java.util.Objects;
