+++ src/main/java/nextstep/ladder/domain/vo/Point.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.vo;
 
 import java.util.HashMap;
 import java.util.Map;
