+++ src/test/java/nextstep/ladder/domain/LadderLinesTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.domain.user.LadderGameUser;
+import nextstep.ladder.domain.vo.Order;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
